Update README.md
This repository contains Task 1 of the Elevate Labs Data Analyst Internship program. The project focuses on data cleaning and preprocessing of the Netflix Movies and TV Shows dataset. The objective was to identify and handle missing values, remove duplicates, standardize inconsistent formats, and prepare a clean, structured dataset ready for analysis. The task was completed using Microsoft Excel, and includes a cleaned dataset along with a detailed README file that explains the steps taken and answers to common data cleaning interview questions.